created useFilter hook
