+++ client/src/pages/MonthlySpendingList.js
@@ -4,6 +4,7 @@ import { useDispatch } from 'react-redux';
 import { Link } from 'react-router-dom';
 
 import { getSpendingByUserId } from '../reducks/spending/operations';
+import useFilter from '../hooks/useFilter';
 
 import MonthFilter from '../components/spending/MonthFilter';
 import SpendingCard from '../components/spending/SpendingCard';
@@ -18,47 +19,17 @@ const MonthlySpendingList = () => {
     '0' + (new Date().getMonth() + 1).toString()
   ).slice(-2)}`;
 
-  const [monthlyTotalSpending, setMonthlyTotalSpending] = useState();
-  const [filterValue, setFilterValue] = useState(initialDate);
-  const [filteredSpendingList, setFilteredSpendingList] = useState([]);
   const userId = useSelector((state) => state.users.uid);
   const spendingList = useSelector((state) => state.spending.spendingList);
+  const [filterValue, setFilterValue] = useState(initialDate);
+
+  const { filteredSpendingList, monthlyTotalSpending } = useFilter(filterValue);
 
   useEffect(() => {
     if (userId) {
       dispatch(getSpendingByUserId(userId));
     }
   }, [dispatch]);
-
-  // filter spending data by month
-  useEffect(() => {
-    if (spendingList && spendingList.length !== 0) {
-      const selectedYear = filterValue.slice(0, 4);
-      const selectedMonth = filterValue.slice(-2);
-
-      setFilteredSpendingList(
-        spendingList.filter((spending) => {
-          return (
-            new Date(spending.createdAt).getFullYear() ===
-              Number(selectedYear) &&
-            new Date(spending.createdAt).getMonth() + 1 ===
-              Number(selectedMonth)
-          );
-        })
-      );
-    }
-  }, [spendingList, filterValue]);
-
-  // get monthly total amount
-  useEffect(() => {
-    if (filteredSpendingList && filteredSpendingList.length !== 0) {
-      const monthlyTotalAmount = filteredSpendingList
-        .map((spending) => spending.amount)
-        .reduce((prev, curr) => prev + curr, 0);
-
-      setMonthlyTotalSpending(monthlyTotalAmount);
-    }
-  }, [filteredSpendingList]);
 
   if (!spendingList || spendingList.length === 0) {
     return (
