+++ client/src/pages/Settings.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Settings = () => {
+  return (
+    <div>Settings</div>
+  )
+}
+
+export default Settings
